requirements.txt
djangorestframework
django-cors-headers
drf-nested-routers
drf-writable-nested
pillow
djoser
pytils

celery
redis

django-cleanup
django-ckeditor

pytils




